24th commit long -> int로 변경 했으나 다시 변경하기! 그리고 일반 로그인 하면 카톡 이미지 뜸 수정 필요! 댓글 작성 구현 완 (image 화면에 출력할 수 있게 수정 필요!)

+++ src/main/java/com/study/myblog/controller/BoardController.java
@@ -28,7 +28,7 @@ public class BoardController {
     }
 
     @GetMapping("/board/{id}")
-    public String findById(@PathVariable Long id, Model model) {
+    public String findById(@PathVariable int id, Model model) {
         model.addAttribute("board", boardService.글상세보기(id));
         return "board/detail";
     }
@@ -40,7 +40,7 @@ public class BoardController {
     }
 
     @GetMapping("/board/{id}/updateForm")
-    public String updateForm(@PathVariable Long id, Model model){
+    public String updateForm(@PathVariable int id, Model model){
         model.addAttribute("board", boardService.글상세보기(id));
         return "board/updateForm";
     }
