+++ src/main/java/com/study/myblog/model/User.java
@@ -18,7 +18,7 @@ import java.sql.Timestamp;
 public class User {
     @Id // Primary key
     @GeneratedValue(strategy = GenerationType.IDENTITY) // 프로젝트에서 연결된 DB의 넘버링 전략을 따라간다.
-    private long id; // 시퀀스, auto_increment
+    private int id; // 시퀀스, auto_increment
     @Column(nullable = false, length = 100, unique = true)
     private String username; // 아이디
     @Column(length = 100)// 넉넉하게 주는 이유는 -> 해시로 변경해서 비밀번호를 암호화 할 것이라서
